+++ salsa20/src/lib.rs
@@ -23,6 +23,21 @@ use cipher::zeroize::{Zeroize, ZeroizeOnDrop};
 mod backends;
 mod xsalsa;
 
+cfg_if::cfg_if! {
+    if #[cfg(any(target_arch = "x86", target_arch = "x86_64"))] {
+        cpufeatures::new!(avx2_cpuid, "avx2");
+        type Tokens = avx2_cpuid::InitToken;
+
+        fn init_tokens() -> Tokens {
+            avx2_cpuid::init()
+        }
+    } else {
+        type Tokens = ();
+
+        fn init_tokens() -> Tokens {}
+    }
+}
+
 pub use xsalsa::{XSalsa8, XSalsa12, XSalsa20, XSalsaCore, hsalsa};
 
 /// Salsa20/8 stream cipher
@@ -56,6 +71,12 @@ const CONSTANTS: [u32; 4] = [0x6170_7865, 0x3320_646e, 0x7962_2d32, 0x6b20_6574]
 pub struct SalsaCore<R: Unsigned> {
     /// Internal state of the core function
     state: [u32; STATE_WORDS],
+    /// CPU target feature tokens
+    #[allow(
+        dead_code,
+        reason = "tokens are held for their type-level proof, used by SIMD backends"
+    )]
+    tokens: Tokens,
     /// Number of rounds to perform
     rounds: PhantomData<R>,
 }
@@ -68,6 +89,7 @@ impl<R: Unsigned> SalsaCore<R> {
     pub fn from_raw_state(state: [u32; STATE_WORDS]) -> Self {
         Self {
             state,
+            tokens: init_tokens(),
             rounds: PhantomData,
         }
     }
@@ -112,6 +134,7 @@ impl<R: Unsigned> KeyIvInit for SalsaCore<R> {
 
         Self {
             state,
+            tokens: init_tokens(),
             rounds: PhantomData,
         }
     }
@@ -124,7 +147,19 @@ impl<R: Unsigned> StreamCipherCore for SalsaCore<R> {
         rem.try_into().ok()
     }
     fn process_with_backend(&mut self, f: impl StreamCipherClosure<BlockSize = Self::BlockSize>) {
-        f.call(&mut backends::soft::Backend(self));
+        cfg_if::cfg_if! {
+            if #[cfg(any(target_arch = "x86", target_arch = "x86_64"))] {
+                if self.tokens.get() {
+                    unsafe {
+                        backends::avx2::inner::<R, _>(&mut self.state, f);
+                    }
+                } else {
+                    f.call(&mut backends::soft::Backend(self));
+                }
+            } else {
+                f.call(&mut backends::soft::Backend(self));
+            }
+        }
     }
 }