+++ salsa20/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- AVX2 intrinsics backend with runtime detection via `cpufeatures` ([#576])
+
 ## 0.11.0 (2026-03-30)
 ### Added
 - SSE2 backend ([#333])
@@ -20,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#333]: https://github.com/RustCrypto/stream-ciphers/pull/333
 [#338]: https://github.com/RustCrypto/stream-ciphers/pull/338
 [#397]: https://github.com/RustCrypto/stream-ciphers/pull/397
+[#576]: https://github.com/RustCrypto/stream-ciphers/pull/576
 
 ## 0.10.2 (2022-02-17)
 ### Added